add dapp support
