+++ dapp/dapp_domain/src/main/java/com/easy/dapp/domain/model/DAppInfo.kt
@@ -1,0 +1,10 @@
+package com.easy.dapp.domain.model
+
+import androidx.annotation.Keep
+
+@Keep
+data class DAppInfo(
+    val url: String,
+    val name: String,
+    val icon: String
+)
