+++ dapp/dapp_presentation/src/main/java/com/easy/dapp/presentation/list/DAppState.kt
@@ -1,0 +1,10 @@
+package com.easy.dapp.presentation.list
+
+import androidx.annotation.Keep
+import com.easy.dapp.domain.model.DAppInfo
+
+@Keep
+data class DAppState(
+    val dApps: Result<List<DAppInfo>>,
+    val isLoading: Boolean
+)
