--- dapp/src/main/java/com/easy/dapp/model/DAppInfo.kt
@@ -1,10 +1,0 @@
-package com.easy.dapp.model
-
-import androidx.annotation.Keep
-
-@Keep
-data class DAppInfo(
-    val url: String,
-    val name: String,
-    val icon: String
-)
