+++ settings.gradle.kts
@@ -38,9 +38,10 @@ includeBuild("dep-version")
 include(":app")
 include(":core")
 include(":core-ui")
-include(":dapp")
 include(":intro")
 include(":settings")
 include(":assets:domain")
 include(":assets:data")
 include(":assets:presentation")
+include(":dapp:dapp_presentation")
+include(":dapp:dapp_domain")
